Reduce size of structure __PHONE_BOOK_ENTRY
To follow methods from former contributor: ryanwang,
I'd like to find out the influence of reducing data size
on cache miss issue. In this program, searching function
is base on comparation between "lastName" and key word of
input, so I comment other unused data in the structure.
The only part modified is the __PHONE_BOOK_ENTRY structure;
functions, "findName" and "append", are the same as those in
phonebook_orig.c.